add advanced example for cornichon-check and fix usability issues

+++ cornichon-check/src/main/scala/com/github/agourlay/cornichon/check/checkModel/ModelRunner.scala
@@ -15,7 +15,7 @@ case class ModelRunner[A, B, C, D, E, F](
 case class Model[A, B, C, D, E, F](
     description: String,
     entryPoint: PropertyN[A, B, C, D, E, F],
-    transitions: Map[PropertyN[A, B, C, D, E, F], List[(Double, PropertyN[A, B, C, D, E, F])]])
+    transitions: Map[PropertyN[A, B, C, D, E, F], List[(Int, PropertyN[A, B, C, D, E, F])]])
 
 object ModelRunner {
